+++ DotNetSample.sln
@@ -99,8 +99,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CustomFormatterWebApp", "sr
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CustomFormatterWebApp.Test", "src\CustomFormatterWebApp.Test\CustomFormatterWebApp.Test.csproj", "{3D4BB4BE-F205-4421-B0CF-E720802CF574}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AzureAdB2cUserManagementConsoleApp", "src\AzureAdB2cUserManagementConsoleApp\AzureAdB2cUserManagementConsoleApp.csproj", "{77026504-A20F-40D7-A267-03D787068F59}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AzureAdB2cMsalConsoleApp", "src\AzureAdB2cMsalConsoleApp\AzureAdB2cMsalConsoleApp.csproj", "{296B47DF-C808-4370-BA76-5DCD5D7056B6}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AzureAdB2cWebApi", "src\AzureAdB2cWebApi\AzureAdB2cWebApi.csproj", "{67B16C25-9632-4D04-9AEB-56DC51A4E844}"
@@ -120,6 +118,8 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenIdConnectWebApp", "src\OpenIdConnectWebApp\OpenIdConnectWebApp.csproj", "{9169C858-A781-458D-BD80-5DAFF0E4BAF9}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RazorPageWebApp", "src\RazorPageWebApp\RazorPageWebApp.csproj", "{AD2A7230-A91F-4AB2-8002-D28A424A4B3C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AzureAdB2cGraphConsoleApp", "src\AzureAdB2cGraphConsoleApp\AzureAdB2cGraphConsoleApp.csproj", "{D6604E96-8541-42B6-9895-11F66186597C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -299,10 +299,6 @@ Global
 		{3D4BB4BE-F205-4421-B0CF-E720802CF574}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3D4BB4BE-F205-4421-B0CF-E720802CF574}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3D4BB4BE-F205-4421-B0CF-E720802CF574}.Release|Any CPU.Build.0 = Release|Any CPU
-		{77026504-A20F-40D7-A267-03D787068F59}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{77026504-A20F-40D7-A267-03D787068F59}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{77026504-A20F-40D7-A267-03D787068F59}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{77026504-A20F-40D7-A267-03D787068F59}.Release|Any CPU.Build.0 = Release|Any CPU
 		{296B47DF-C808-4370-BA76-5DCD5D7056B6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{296B47DF-C808-4370-BA76-5DCD5D7056B6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{296B47DF-C808-4370-BA76-5DCD5D7056B6}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -343,6 +339,10 @@ Global
 		{AD2A7230-A91F-4AB2-8002-D28A424A4B3C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AD2A7230-A91F-4AB2-8002-D28A424A4B3C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{AD2A7230-A91F-4AB2-8002-D28A424A4B3C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D6604E96-8541-42B6-9895-11F66186597C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D6604E96-8541-42B6-9895-11F66186597C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D6604E96-8541-42B6-9895-11F66186597C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D6604E96-8541-42B6-9895-11F66186597C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -388,7 +388,6 @@ Global
 		{F9802D27-7A90-40D3-ADC3-3D50B39F7206} = {512A5E50-BE18-4761-BA83-E56F4EC43B9E}
 		{AF0FE8C6-126F-446F-BDAB-6C1E6EE98815} = {512A5E50-BE18-4761-BA83-E56F4EC43B9E}
 		{3D4BB4BE-F205-4421-B0CF-E720802CF574} = {512A5E50-BE18-4761-BA83-E56F4EC43B9E}
-		{77026504-A20F-40D7-A267-03D787068F59} = {798EF2B5-D82E-49ED-8EDE-BD3B20040D59}
 		{296B47DF-C808-4370-BA76-5DCD5D7056B6} = {798EF2B5-D82E-49ED-8EDE-BD3B20040D59}
 		{67B16C25-9632-4D04-9AEB-56DC51A4E844} = {512A5E50-BE18-4761-BA83-E56F4EC43B9E}
 		{2EFD1B28-15EB-4EE7-BD95-808C157BCF35} = {512A5E50-BE18-4761-BA83-E56F4EC43B9E}
@@ -399,6 +398,7 @@ Global
 		{39511B3C-80D5-4A01-8268-C84E884EE17F} = {512A5E50-BE18-4761-BA83-E56F4EC43B9E}
 		{9169C858-A781-458D-BD80-5DAFF0E4BAF9} = {512A5E50-BE18-4761-BA83-E56F4EC43B9E}
 		{AD2A7230-A91F-4AB2-8002-D28A424A4B3C} = {512A5E50-BE18-4761-BA83-E56F4EC43B9E}
+		{D6604E96-8541-42B6-9895-11F66186597C} = {798EF2B5-D82E-49ED-8EDE-BD3B20040D59}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {D8A5A7FA-225F-46ED-9227-A5AB57BDF755}
