+++ DotNetSampleConsole.slnf
@@ -2,8 +2,8 @@
   "solution": {
     "path": "DotNetSample.sln",
     "projects": [
+      "src\\AzureAdB2cGraphConsoleApp\\AzureAdB2cGraphConsoleApp.csproj",
       "src\\AzureAdB2cMsalConsoleApp\\AzureAdB2cMsalConsoleApp.csproj",
-      "src\\AzureAdB2cUserManagementConsoleApp\\AzureAdB2cUserManagementConsoleApp.csproj",
       "src\\AzureCacheForRedisConsoleApp\\AzureCacheForRedisConsoleApp.csproj",
       "src\\AzureCosmosDbConsoleApp\\AzureCosmosDbConsoleApp.csproj",
       "src\\AzureStorageConsoleApp\\AzureStorageConsoleApp.csproj",
