+++ src/AzureAdB2cGraphConsoleApp/CustomAttributeHelper.cs
@@ -1,0 +1,7 @@
+namespace AzureAdB2cGraphConsoleApp;
+
+public class CustomAttributeHelper(string extensionAppClientId) {
+	private readonly string _extensionAppClientId = extensionAppClientId.Replace("-", "");
+
+	public string GetFullName(string attributeName) => $"extension_{_extensionAppClientId}_{attributeName}";
+}
