+++ src/AzureAdB2cGraphConsoleApp/CustomAttributeNames.cs
@@ -1,0 +1,5 @@
+namespace AzureAdB2cGraphConsoleApp;
+
+public static class CustomAttributeNames {
+	public const string TestNumber = nameof(TestNumber);
+}
