+++ src/AzureAdB2cGraphConsoleApp/appsettings.json
@@ -1,0 +1,17 @@
+{
+	// テナント名
+	"TenantId": "",
+	// クライアントID
+	"ClientId": "",
+	// クライアントシークレット
+	"ClientSecret": "",
+	// 「b2c-extensions-app. Do not modify. Used by AADB2C for storing user data.」というアプリのクライアントID
+	"ExtensionAppClientId": "",
+	"Logging": {
+		"LogLevel": {
+			"Default": "Information",
+			"Microsoft": "Warning",
+			"Microsoft.Hosting.Lifetime": "Information"
+		}
+	}
+}
