--- src/AzureAdB2cUserManagementConsoleApp/AzureAdB2cUserManagementConsoleApp.csproj
@@ -1,20 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net8.0</TargetFramework>
-    <UserSecretsId>ca51f549-61fd-45fe-9eea-e4088aa7249a</UserSecretsId>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Azure.Identity" Version="1.11.3" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.UserSecrets" Version="8.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="8.0.0" />
-    <PackageReference Include="Microsoft.Graph" Version="5.53.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\SampleLib\SampleLib.csproj" />
-  </ItemGroup>
-
-</Project>
