--- src/AzureAdB2cUserManagementConsoleApp/CustomAttributeHelper.cs
@@ -1,7 +1,0 @@
-namespace AzureAdB2cUserManagementConsoleApp;
-
-public class CustomAttributeHelper(string extensionAppClientId) {
-	private readonly string _extensionAppClientId = extensionAppClientId.Replace("-", "");
-
-	public string GetFullName(string attributeName) => $"extension_{_extensionAppClientId}_{attributeName}";
-}
