--- src/AzureAdB2cUserManagementConsoleApp/CustomAttributeNames.cs
@@ -1,5 +1,0 @@
-namespace AzureAdB2cUserManagementConsoleApp;
-
-public static class CustomAttributeNames {
-	public const string TestNumber = nameof(TestNumber);
-}
