--- src/AzureAdB2cUserManagementConsoleApp/Properties/launchSettings.json
@@ -1,10 +1,0 @@
-{
-  "profiles": {
-    "AzureAdB2cUserManagementConsoleApp": {
-      "commandName": "Project",
-      "environmentVariables": {
-        "DOTNET_ENVIRONMENT": "Development"
-      }
-    }
-  }
-}
