--- src/AzureAdB2cUserManagementConsoleApp/appsettings.json
@@ -1,17 +1,0 @@
-{
-	// テナント名
-	"TenantId": "",
-	// クライアントID
-	"ClientId": "",
-	// クライアントシークレット
-	"ClientSecret": "",
-	// 「b2c-extensions-app. Do not modify. Used by AADB2C for storing user data.」というアプリのクライアントID
-	"ExtensionAppClientId": "",
-	"Logging": {
-		"LogLevel": {
-			"Default": "Information",
-			"Microsoft": "Warning",
-			"Microsoft.Hosting.Lifetime": "Information"
-		}
-	}
-}
